pyproject.toml
[project]
name = "federated-task-arithmetic"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "pymupdf>=1.22.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.2.0",
    "timm>=0.9.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.scripts]
fl-train = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
